:card_file_box: Mise a jour des cles primaires et etrangere des tables concernants le forum

+++ Modules/Forum/Database/Migrations/2017_03_14_180530_create_replies_table.php
@@ -14,9 +14,9 @@ class CreateRepliesTable extends Migration
     public function up()
     {
         Schema::create('replies', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('thread_id');
-            $table->integer('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('thread_id');
+            $table->unsignedBigInteger('user_id');
             $table->text('body');
             $table->timestamps();
         });
