+++ Modules/Forum/Database/Migrations/2017_03_30_161938_create_favorites_table.php
@@ -14,9 +14,9 @@ class CreateFavoritesTable extends Migration
     public function up()
     {
         Schema::create('favorites', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('favorited_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('favorited_id');
             $table->string('favorited_type', 50);
             $table->timestamps();
