+++ Modules/Forum/Database/Migrations/2017_04_21_162429_create_activities_table.php
@@ -14,9 +14,9 @@ class CreateActivitiesTable extends Migration
     public function up()
     {
         Schema::create('activities', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id')->index();
-            $table->unsignedInteger('subject_id')->index();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedBigInteger('subject_id')->index();
             $table->string('subject_type', 50);
             $table->string('type', 50);
             $table->timestamps();
