+++ Modules/Forum/Database/Migrations/2017_05_19_141013_create_thread_subscriptions_table.php
@@ -14,9 +14,9 @@ class CreateThreadSubscriptionsTable extends Migration
     public function up()
     {
         Schema::create('thread_subscriptions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('thread_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('thread_id');
             $table->timestamps();
             $table->unique(['user_id', 'thread_id']);
